modularized queue functions
